ns-by-example Extending NS : Add New Application and Agent http://nile.wpi.edu/NS/new_app_agent.html
Download "mm-app.h", "mm-app.cc", "udp-mm.h" and "udp-mm.cc") to the "ns-2" directory.
Make sure you registered the new application header by modifying "packet.h" and "ns-packet.tcl" as shown in Figure 24 (a) and (b).
Add supportMM() and enableMM() methods to the "Agent" class in "agent.h" as shown in Figure 26 (a).
Add recv_msg() method to the "Application" class in "app.h" as shown in Figure 26 (b).
Set default values for the newly introduced configurable parameters in "ns-default.tcl" as described in Figure 27. Be SURE to complete this last step.
 Otherwise, all five-scale rates are initialized to zero unless specified in the input simulation script (i.e., the test simulation script given below will not transmit any frames).